Changed spelling of receive_money method and added change_balance setter
Card Number: 7
The `receive_money(amount)` method was previously misspelled `recieve_money(amount)` and was corrected to prevent future bugs.

The `change_balance(new_total)` method (change_balance(int) : void) was added to function as a setter for the private member data `__money`, which is the total monetary balance of the player. This method was primarily added to accommodate the implementation of the transactional methods in the bank class.