[27] and [31] Defined the Deed Class and Matched with json
Created a Deeds hierarchy that mirrors the tiles.py hierarchy and matches the tile.json, properties.json, utilities.json, and railroads.json. The streets, railroads, and utilities deeds all derive from a base deeds class (which will be abstract in the future) which carries the universal attributes of each type of deed. In addition, the Utilities deed was temporarily given the attributes `rent_multiplier_one` and `rent_multiplier_two`, which encompasses the rent multiplier when either one or both utilities are owned. These are hard-coded multipliers for now and can be changed to also read from the file after their possible addition is discussed and finalized.