Added tiles.json formatting reference
This file contains the formatting of each type of tile in the tiles.json file. The various types of tiles are:
- special-corner
- special-card
- special-tax
- utility
- railroad
- street

The formatting for each type is unique, with some similarities between types (e.g. Utilities have the same image generation as special-cards, but the price display of other properties.) This file is useful for those that don't want to scroll through the entire tiles.json file to locate the formatting of one type of tile.