Genericized moveTo attribute in community_chest.json
Changed moveTo from string type (giving the name of the location) to int type (giving the index of the location) to make future regional or otherwise customized variants of Monopoly (which may change the name of the location to accommodate the region) easier to program. This also reflects the variable typing of Chance cards with the moveTo attribute.