[22] Filled out chance.json card attributes
The majority of Chance cards have the Movement type and thus have the moveTo attribute (int type referring to index of target location.) Uniquely, there exists only one "reverse movement" card in the game, in which the player is instructed to move counter-clockwise. The moveTo attribute was set to a negative value to accommodate this difference, and would thus require implementation to be 'current_location + moveTo' rather than 'current_location = moveTo'. And yes, most Chance cards are printed in all caps, although this can be changed later.